pyproject.toml
[project]
name = "cnn-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grad-cam>=1.5.5",
    "gradio>=6.3.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
